+++ src/LayersControl.js
@@ -164,7 +164,11 @@ export default class LayersControl extends MapControl {
     const children = Children.map(this.props.children, child => {
       return child ? cloneElement(child, this.controlProps) : null
     })
-    return <div style={{ display: 'none' }}>{children}</div>
+    return (
+      <div style={{ display: 'none' }}>
+        {children}
+      </div>
+    )
   }
 }